requirements.txt
astroid==2.1.0
autopep8==1.4.3
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
coveralls==1.5.1
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-extensions==2.1.2
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
docopt==0.6.2
facebook-sdk==3.1.0
future==0.17.1
google-api-python-client==1.7.6
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.9.0
httplib2==0.12.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
oauthlib==2.1.0
openapi-codec==1.3.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
PyJWT==1.4.2
pylint==2.2.1
python-twitter==3.5
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
rsa==4.0
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
